refactor(extract): move lazarus_package, slnx, csproj to extractors/ (verbatim)
Moves three bespoke XML extractors out of the extract.py monolith into
their own modules per the MIGRATION.md playbook (#1212). Shared XML
safety helpers (_PROJECT_XML_MAX_BYTES, _project_xml_is_safe) promoted
to extractors/base.py, unblocking future migrations of xaml and any
other XML-based extractor. Facade re-exports and registry entries added;
zero test changes outside test_extractors_registry.py.
