+++ graphify/extract.py
@@ -147,6 +147,10 @@ from graphify.extractors.pascal import _PAS_BEGIN_END_TOKEN_RE, _PAS_CALL_RE, _P
 from graphify.extractors.objc import _objc_local_var_types, extract_objc  # noqa: E402,F401
 
 from graphify.extractors.julia import extract_julia  # noqa: E402,F401
+from graphify.extractors.lazarus_package import extract_lazarus_package  # noqa: E402,F401
+from graphify.extractors.slnx import extract_slnx  # noqa: E402,F401
+from graphify.extractors.csproj import extract_csproj  # noqa: E402,F401
+from graphify.extractors.base import _PROJECT_XML_MAX_BYTES, _project_xml_is_safe  # noqa: E402,F401
 
 _RECURSION_LIMIT = 10_000
 
@@ -3759,120 +3763,6 @@ register_language_resolver(
 # ── Pascal / Delphi extractor ─────────────────────────────────────────────────
 
 
-# Size cap for project XML files we parse with stdlib ElementTree.
-# Real .csproj/.fsproj/.vbproj/.lpk files are well under 2 MiB; anything
-# larger is either malformed or hostile.
-_PROJECT_XML_MAX_BYTES = 2 * 1024 * 1024
-
-
-def _project_xml_is_safe(src: bytes) -> bool:
-    """Reject XML that declares DTDs or entities.
-
-    Stdlib ``xml.etree.ElementTree`` does not cap entity expansion, so a
-    crafted project file could trigger a billion-laughs style DoS. External
-    entity resolution is already disabled by pyexpat defaults, but rejecting
-    ``<!DOCTYPE`` / ``<!ENTITY`` outright is defense in depth.
-
-    Legitimate MSBuild and Lazarus package files never contain a DOCTYPE
-    or ENTITY declaration, so this is a zero-false-positive screen.
-    """
-    # Only the prolog can hold a DTD/internal subset, but be conservative
-    # and scan the full byte range -- these formats use ASCII tags so a
-    # case-insensitive substring match is sufficient.
-    lowered = src.lower()
-    return b"<!doctype" not in lowered and b"<!entity" not in lowered
-
-
-def extract_lazarus_package(path: Path) -> dict:
-    """Extract package metadata from Lazarus .lpk package files (XML format).
-
-    .lpk is an XML file listing the package name, required dependencies,
-    and the Pascal units that belong to the package.
-
-    Produces nodes for:
-    - The package file itself
-    - The package (by name)
-    - Each required package (dependency)
-    - Each listed unit file (resolved to path-based IDs where possible)
-
-    Produces edges for:
-    - file --contains--> package
-    - package --imports--> required dependency (context: "import")
-    - package --contains--> listed unit
-    """
-    try:
-        import xml.etree.ElementTree as ET
-        src = path.read_bytes()
-    except OSError as e:
-        return {"nodes": [], "edges": [], "error": str(e)}
-
-    if len(src) > _PROJECT_XML_MAX_BYTES:
-        return {"nodes": [], "edges": [], "error": "package file too large"}
-    if not _project_xml_is_safe(src):
-        return {"nodes": [], "edges": [],
-                "error": "refusing XML with DOCTYPE/ENTITY declaration"}
-
-    try:
-        xml_root = ET.fromstring(src)
-    except Exception as e:
-        return {"nodes": [], "edges": [], "error": str(e)}
-
-    str_path = str(path)
-    stem = _file_stem(path)
-    nodes: list[dict] = []
-    edges: list[dict] = []
-    seen_ids: set[str] = set()
-
-    def add_node(nid: str, label: str) -> None:
-        if nid not in seen_ids:
-            seen_ids.add(nid)
-            nodes.append({
-                "id": nid, "label": label, "file_type": "code",
-                "source_file": str_path, "source_location": "L1",
-            })
-
-    def add_edge(src: str, tgt: str, relation: str, context: str | None = None) -> None:
-        edge: dict[str, Any] = {
-            "source": src, "target": tgt, "relation": relation,
-            "confidence": "EXTRACTED", "source_file": str_path,
-            "source_location": "L1", "weight": 1.0,
-        }
-        if context:
-            edge["context"] = context
-        edges.append(edge)
-
-    file_nid = _make_id(str(path))
-    add_node(file_nid, path.name)
-
-    name_elem = xml_root.find(".//Package/Name")
-    pkg_name = name_elem.get("Value") if name_elem is not None else path.stem
-    pkg_nid = _make_id(stem, pkg_name)
-    add_node(pkg_nid, pkg_name)
-    add_edge(file_nid, pkg_nid, "contains")
-
-    # Required packages → imports edges
-    for item in xml_root.findall(".//RequiredPkgs/"):
-        dep_elem = item.find("PackageName")
-        if dep_elem is not None:
-            dep_name = dep_elem.get("Value", "")
-            if dep_name:
-                dep_nid = _make_id(dep_name)
-                add_node(dep_nid, dep_name)
-                add_edge(pkg_nid, dep_nid, "imports", context="import")
-
-    # Listed units → contains edges, resolved to path-based IDs where possible
-    for item in xml_root.findall(".//Files/"):
-        unit_elem = item.find("UnitName")
-        if unit_elem is not None:
-            unit_name = unit_elem.get("Value", "")
-            if unit_name:
-                unit_nid = _pascal_resolve_unit(path, unit_name)
-                add_node(unit_nid, unit_name)
-                add_edge(pkg_nid, unit_nid, "contains")
-
-    return {"nodes": nodes, "edges": edges, "input_tokens": 0, "output_tokens": 0}
-
-
 # ── Main extract and collect_files ────────────────────────────────────────────
 
 
@@ -3895,206 +3785,6 @@ def _check_tree_sitter_version() -> None:
 
 
 # ── .NET project files (.sln, .slnx, .csproj, .razor) ───────────────────────
-
-
-def extract_slnx(path: Path) -> dict:
-    """Extract projects and inter-project dependencies from a .slnx file.
-
-    .slnx is the XML-based replacement for the legacy .sln format. Projects
-    are listed as ``<Project Path="..."/>`` elements (optionally nested inside
-    ``<Folder>`` elements) and build-order dependencies as ``<BuildDependency
-    Project="..."/>`` children. Unlike .sln there are no GUIDs -- projects are
-    identified by their path.
-    """
-    import xml.etree.ElementTree as ET
-
-    try:
-        src = path.read_bytes()
-    except OSError:
-        return {"nodes": [], "edges": [], "error": f"cannot read {path}"}
-
-    if len(src) > _PROJECT_XML_MAX_BYTES:
-        return {"nodes": [], "edges": [], "error": "project file too large"}
-    if not _project_xml_is_safe(src):
-        return {"nodes": [], "edges": [],
-                "error": "refusing XML with DOCTYPE/ENTITY declaration"}
-
-    try:
-        tree = ET.fromstring(src)
-    except ET.ParseError as e:
-        return {"nodes": [], "edges": [], "error": f"XML parse error: {e}"}
-
-    file_nid = _make_id(str(path))
-    str_path = str(path)
-    nodes: list[dict] = [{"id": file_nid, "label": path.name, "file_type": "code",
-                          "source_file": str_path, "source_location": None}]
-    edges: list[dict] = []
-    seen_ids: set[str] = set()
-    seen_ids.add(file_nid)
-
-    ns = ""
-    if tree.tag.startswith("{"):
-        ns = tree.tag.split("}")[0] + "}"
-
-    def _resolve(proj_path: str) -> str:
-        proj_path = proj_path.replace("\\", "/")
-        try:
-            return str((path.parent / proj_path).resolve())
-        except Exception:
-            return proj_path
-
-    # First pass: collect projects (anywhere in the tree, incl. <Folder>).
-    project_nids: set[str] = set()
-    for proj in tree.iter(f"{ns}Project"):
-        proj_path = proj.get("Path")
-        if not proj_path:
-            continue
-        abs_proj = _resolve(proj_path)
-        proj_nid = _make_id(abs_proj)
-        if proj_nid and proj_nid not in seen_ids:
-            seen_ids.add(proj_nid)
-            label = Path(proj_path).stem
-            nodes.append({"id": proj_nid, "label": label,
-                          "file_type": "code", "source_file": abs_proj,
-                          "source_location": None})
-            edges.append({"source": file_nid, "target": proj_nid,
-                          "relation": "contains", "confidence": "EXTRACTED",
-                          "source_file": str_path, "weight": 1.0})
-        if proj_nid:
-            project_nids.add(proj_nid)
-
-    # Second pass: build-order dependencies between known projects.
-    for proj in tree.iter(f"{ns}Project"):
-        proj_path = proj.get("Path")
-        if not proj_path:
-            continue
-        from_nid = _make_id(_resolve(proj_path))
-        for dep in proj.iter(f"{ns}BuildDependency"):
-            dep_path = dep.get("Project")
-            if not dep_path:
-                continue
-            to_nid = _make_id(_resolve(dep_path))
-            if (from_nid and to_nid and from_nid != to_nid
-                    and to_nid in project_nids):
-                edges.append({"source": from_nid, "target": to_nid,
-                              "relation": "imports", "confidence": "EXTRACTED",
-                              "source_file": str_path, "weight": 1.0})
-
-    return {"nodes": nodes, "edges": edges}
-
-
-def extract_csproj(path: Path) -> dict:
-    """Extract packages, project refs, and target framework from a .csproj/.fsproj/.vbproj."""
-    import xml.etree.ElementTree as ET
-
-    try:
-        src = path.read_bytes()
-    except OSError:
-        return {"nodes": [], "edges": [], "error": f"cannot read {path}"}
-
-    if len(src) > _PROJECT_XML_MAX_BYTES:
-        return {"nodes": [], "edges": [], "error": "project file too large"}
-    if not _project_xml_is_safe(src):
-        return {"nodes": [], "edges": [],
-                "error": "refusing XML with DOCTYPE/ENTITY declaration"}
-
-    try:
-        tree = ET.fromstring(src)
-    except ET.ParseError as e:
-        return {"nodes": [], "edges": [], "error": f"XML parse error: {e}"}
-
-    file_nid = _make_id(str(path))
-    str_path = str(path)
-    nodes: list[dict] = [{"id": file_nid, "label": path.name, "file_type": "code",
-                          "source_file": str_path, "source_location": None}]
-    edges: list[dict] = []
-    seen_ids: set[str] = set()
-    seen_ids.add(file_nid)
-
-    ns = ""
-    root_tag = tree.tag
-    if root_tag.startswith("{"):
-        ns = root_tag.split("}")[0] + "}"
-
-    def find_all(tag: str):
-        return tree.iter(f"{ns}{tag}")
-
-    for tf in find_all("TargetFramework"):
-        if tf.text:
-            fw_nid = _make_id("framework", tf.text.strip())
-            if fw_nid and fw_nid not in seen_ids:
-                seen_ids.add(fw_nid)
-                nodes.append({"id": fw_nid, "label": tf.text.strip(),
-                              "file_type": "concept", "source_file": str_path,
-                              "source_location": None})
-                edges.append({"source": file_nid, "target": fw_nid,
-                              "relation": "references", "confidence": "EXTRACTED",
-                              "source_file": str_path, "weight": 1.0})
-
-    for tf in find_all("TargetFrameworks"):
-        if tf.text:
-            for fw in tf.text.strip().split(";"):
-                fw = fw.strip()
-                if fw:
-                    fw_nid = _make_id("framework", fw)
-                    if fw_nid and fw_nid not in seen_ids:
-                        seen_ids.add(fw_nid)
-                        nodes.append({"id": fw_nid, "label": fw,
-                                      "file_type": "concept", "source_file": str_path,
-                                      "source_location": None})
-                        edges.append({"source": file_nid, "target": fw_nid,
-                                      "relation": "references", "confidence": "EXTRACTED",
-                                      "source_file": str_path, "weight": 1.0})
-
-    for pkg in find_all("PackageReference"):
-        name = pkg.get("Include") or pkg.get("include") or ""
-        version = pkg.get("Version") or pkg.get("version") or ""
-        if not name:
-            continue
-        pkg_nid = _make_id("nuget", name)
-        label = f"{name} ({version})" if version else name
-        if pkg_nid and pkg_nid not in seen_ids:
-            seen_ids.add(pkg_nid)
-            nodes.append({"id": pkg_nid, "label": label,
-                          "file_type": "code", "source_file": str_path,
-                          "source_location": None})
-        edges.append({"source": file_nid, "target": pkg_nid,
-                      "relation": "imports", "confidence": "EXTRACTED",
-                      "source_file": str_path, "weight": 1.0})
-
-    for proj in find_all("ProjectReference"):
-        ref_path = proj.get("Include") or proj.get("include") or ""
-        if not ref_path:
-            continue
-        ref_path_norm = ref_path.replace("\\", "/")
-        try:
-            abs_ref = str((path.parent / ref_path_norm).resolve())
-        except Exception:
-            abs_ref = ref_path_norm
-        proj_nid = _make_id(abs_ref)
-        if proj_nid and proj_nid not in seen_ids:
-            seen_ids.add(proj_nid)
-            proj_label = Path(ref_path_norm).name
-            nodes.append({"id": proj_nid, "label": proj_label,
-                          "file_type": "code", "source_file": abs_ref,
-                          "source_location": None})
-        edges.append({"source": file_nid, "target": proj_nid,
-                      "relation": "imports", "confidence": "EXTRACTED",
-                      "source_file": str_path, "weight": 1.0})
-
-    sdk = tree.get("Sdk") or ""
-    if sdk:
-        sdk_nid = _make_id("sdk", sdk)
-        if sdk_nid and sdk_nid not in seen_ids:
-            seen_ids.add(sdk_nid)
-            nodes.append({"id": sdk_nid, "label": sdk,
-                          "file_type": "concept", "source_file": str_path,
-                          "source_location": None})
-            edges.append({"source": file_nid, "target": sdk_nid,
-                          "relation": "references", "confidence": "EXTRACTED",
-                          "source_file": str_path, "weight": 1.0})
-
-    return {"nodes": nodes, "edges": edges}
 
 
 def _xml_local_name(name: str) -> str:
