+++ graphify/extractors/__init__.py
@@ -13,6 +13,7 @@ from typing import Callable
 from graphify.extractors.apex import extract_apex
 from graphify.extractors.bash import extract_bash
 from graphify.extractors.blade import extract_blade
+from graphify.extractors.csproj import extract_csproj
 from graphify.extractors.dart import extract_dart
 from graphify.extractors.dm import extract_dm, extract_dmf, extract_dmi, extract_dmm
 from graphify.extractors.elixir import extract_elixir
@@ -20,6 +21,7 @@ from graphify.extractors.fortran import extract_fortran
 from graphify.extractors.go import extract_go
 from graphify.extractors.json_config import extract_json
 from graphify.extractors.julia import extract_julia
+from graphify.extractors.lazarus_package import extract_lazarus_package
 from graphify.extractors.markdown import extract_markdown
 from graphify.extractors.objc import extract_objc
 from graphify.extractors.pascal import extract_pascal
@@ -28,6 +30,7 @@ from graphify.extractors.powershell import extract_powershell, extract_powershel
 from graphify.extractors.razor import extract_razor
 from graphify.extractors.rust import extract_rust
 from graphify.extractors.sln import extract_sln
+from graphify.extractors.slnx import extract_slnx
 from graphify.extractors.sql import extract_sql
 from graphify.extractors.terraform import extract_terraform
 from graphify.extractors.verilog import extract_verilog
@@ -36,6 +39,7 @@ from graphify.extractors.zig import extract_zig
 LANGUAGE_EXTRACTORS: dict[str, Callable[[Path], dict]] = {
     "apex": extract_apex,
     "bash": extract_bash,
+    "csproj": extract_csproj,
     "blade": extract_blade,
     "dart": extract_dart,
     "delphi_form": extract_delphi_form,
@@ -49,6 +53,7 @@ LANGUAGE_EXTRACTORS: dict[str, Callable[[Path], dict]] = {
     "json": extract_json,
     "julia": extract_julia,
     "lazarus_form": extract_lazarus_form,
+    "lazarus_package": extract_lazarus_package,
     "markdown": extract_markdown,
     "objc": extract_objc,
     "pascal": extract_pascal,
@@ -57,6 +62,7 @@ LANGUAGE_EXTRACTORS: dict[str, Callable[[Path], dict]] = {
     "razor": extract_razor,
     "rust": extract_rust,
     "sln": extract_sln,
+    "slnx": extract_slnx,
     "sql": extract_sql,
     "terraform": extract_terraform,
     "verilog": extract_verilog,
